working post middleware
